Better pause, reformat isotopes, isotope dropdown
the simulation keeps track of how much time has passed by cheing when the simulation started and what time it is. Pause and unpause work by setting the start time or the 'pause-time' accordingly. simplified and corrected the isotopes.txt file. Parse the file in and create a drop down with the information.